+++ tree-sitter/src/TreeSitter/GenerateSyntax.hs
@@ -16,6 +16,7 @@ module TreeSitter.GenerateSyntax
 
 import Data.Char
 import Language.Haskell.TH as TH
+import Language.Haskell.TH.Syntax as TH
 import Data.HashSet (HashSet)
 import TreeSitter.Deserialize (Datatype (..), DatatypeName (..), Field (..), Children(..), Required (..), Type (..), Named (..), Multiple (..))
 import Data.List.NonEmpty (NonEmpty (..))
@@ -27,7 +28,6 @@ import GHC.Generics hiding (Constructor, Datatype)
 import Foreign.Ptr
 import qualified TreeSitter.Language as TS
 import Foreign.C.String
-import Data.Proxy
 import Data.Aeson hiding (String)
 import System.Directory
 import System.FilePath.Posix
@@ -36,7 +36,11 @@ import TreeSitter.Token
 import TreeSitter.Symbol (escapeOperatorPunctuation)
 
 
--- Auto-generate Haskell datatypes from node-types.json
+-- | Derive Haskell datatypes from a language and its @node-types.json@ file.
+--
+-- Datatypes will be generated according to the specification in the @node-types.json@ file, with anonymous leaf types defined as synonyms for the 'Token' datatype.
+--
+-- Any datatypes among the node types which have already been defined in the module where the splice is run will be skipped, allowing customization of the representation of parts of the tree. Note that this should be used sparingly, as it imposes extra maintenance burden, particularly when the grammar is changed. This may be used to e.g. parse literals into Haskell equivalents (e.g. parsing the textual contents of integer literals into 'Integer's), and may require defining 'TS.UnmarshalAnn' or 'TS.SymbolMatching' instances for (parts of) the custom datatypes, depending on where and how the datatype occurs in the generated tree, in addition to the usual 'Foldable', 'Functor', etc. instances provided for generated datatypes.
 astDeclarationsForLanguage :: Ptr TS.Language -> FilePath -> Q [Dec]
 astDeclarationsForLanguage language filePath = do
   _ <- TS.addDependentFileRelative filePath
@@ -49,13 +53,13 @@ astDeclarationsForLanguage language filePath = do
 
 -- Auto-generate Haskell datatypes for sums, products and leaf types
 syntaxDatatype :: Ptr TS.Language -> Datatype -> Q [Dec]
-syntaxDatatype language datatype = do
+syntaxDatatype language datatype = skipDefined $ do
   typeParameterName <- newName "a"
   case datatype of
-    SumType (DatatypeName datatypeName) _ subtypes -> do
-      cons <- traverse (constructorForSumChoice datatypeName typeParameterName) subtypes
-      result <- symbolMatchingInstanceForSums name subtypes
-      pure $ generatedDatatype name cons typeParameterName:result
+    SumType (DatatypeName _) _ subtypes -> do
+      types' <- fieldTypesToNestedSum subtypes
+      con <- normalC name [TH.bangType strictness (pure types' `appT` varT typeParameterName)]
+      pure [NewtypeD [] name [PlainTV typeParameterName] Nothing con [deriveGN, deriveStockClause, deriveAnyClassClause]]
     ProductType (DatatypeName datatypeName) _ children fields -> do
       con <- ctorForProductType datatypeName typeParameterName children fields
       result <- symbolMatchingInstance language name datatypeName
@@ -68,9 +72,16 @@ syntaxDatatype language datatype = do
       result <- symbolMatchingInstance language name datatypeName
       pure $ generatedDatatype name [con] typeParameterName:result
   where
-    name = toName (datatypeNameStatus datatype) (getDatatypeName (TreeSitter.Deserialize.datatypeName datatype))
-    deriveClause = [ DerivClause Nothing [ ConT ''TS.Unmarshal, ConT ''Eq, ConT ''Ord, ConT ''Show, ConT ''Generic, ConT ''Foldable, ConT ''Functor, ConT ''Traversable, ConT ''Generic1] ]
-    generatedDatatype name cons typeParameterName = DataD [] name [PlainTV typeParameterName] Nothing cons deriveClause
+    -- Skip generating datatypes that have already been defined (overridden) in the module where the splice is running.
+    skipDefined m = do
+      isLocal <- lookupTypeName nameStr >>= maybe (pure False) isLocalName
+      if isLocal then pure [] else m
+    name = mkName nameStr
+    nameStr = toNameString (datatypeNameStatus datatype) (getDatatypeName (TreeSitter.Deserialize.datatypeName datatype))
+    deriveStockClause = DerivClause (Just StockStrategy) [ ConT ''Eq, ConT ''Ord, ConT ''Show, ConT ''Generic, ConT ''Foldable, ConT ''Functor, ConT ''Traversable, ConT ''Generic1]
+    deriveAnyClassClause = DerivClause (Just AnyclassStrategy) [ConT ''TS.Unmarshal]
+    deriveGN = DerivClause (Just NewtypeStrategy) [ConT ''TS.SymbolMatching]
+    generatedDatatype name cons typeParameterName = DataD [] name [PlainTV typeParameterName] Nothing cons [deriveStockClause, deriveAnyClassClause]
 
 
 -- | Create TH-generated SymbolMatching instances for sums, products, leaves
@@ -82,21 +93,6 @@ symbolMatchingInstance language name str = do
       showFailure _ node = "Expected " <> $(litE (stringL (show name))) <> " but got " <> show (TS.fromTSSymbol (nodeSymbol node) :: $(conT (mkName "Grammar.Grammar")))
       symbolMatch _ node = TS.fromTSSymbol (nodeSymbol node) == $(conE (mkName $ "Grammar." <> TS.symbolToName tsSymbolType str))|]
 
-symbolMatchingInstanceForSums :: Name -> [TreeSitter.Deserialize.Type] -> Q [Dec]
-symbolMatchingInstanceForSums name subtypes =
-  [d|instance TS.SymbolMatching $(conT name) where
-      showFailure _ node = "Expected " <> $(litE (stringL (show (map extractn subtypes)))) <> " but got " <> show (TS.fromTSSymbol (nodeSymbol node) :: $(conT (mkName "Grammar.Grammar")))
-      symbolMatch _ = $(foldr1 mkOr (perMkType `map` subtypes)) |]
-  where perMkType (MkType (DatatypeName n) named) = [e|TS.symbolMatch (Proxy :: Proxy $(conT (toName named n))) |]
-        mkOr lhs rhs = [e| (||) <$> $(lhs) <*> $(rhs) |]
-        extractn (MkType (DatatypeName n) Named) = toCamelCase n
-        extractn (MkType (DatatypeName n) Anonymous) = "Anonymous" <> toCamelCase n
-
-
--- | Append string with constructor name (ex., @IfStatementStatement IfStatement@)
-constructorForSumChoice :: String -> Name -> TreeSitter.Deserialize.Type -> Q Con
-constructorForSumChoice str typeParameterName (MkType (DatatypeName n) named) = normalC (toName named (n ++ str)) [child]
-  where child = TH.bangType strictness (appT (conT (toName named n)) (varT typeParameterName))
 
 -- | Build Q Constructor for product types (nodes with fields)
 ctorForProductType :: String -> Name -> Maybe Children -> [(String, Field)] -> Q Con
@@ -106,7 +102,7 @@ ctorForProductType constructorName typeParameterName children fields = ctorForTy
   fieldList = map (fmap toType) fields
   childList = toList $ fmap toTypeChild children
   toType (MkField required fieldTypes mult) =
-    let ftypes = fieldTypesToNestedEither fieldTypes typeParameterName
+    let ftypes = fieldTypesToNestedSum fieldTypes `appT` varT typeParameterName
     in case (required, mult) of
       (Required, Multiple) -> appT (conT ''NonEmpty) ftypes
       (Required, Single) -> ftypes
@@ -117,8 +113,8 @@ ctorForProductType constructorName typeParameterName children fields = ctorForTy
 -- | Build Q Constructor for leaf types (nodes with no fields or subtypes)
 ctorForLeafType :: DatatypeName -> Name -> Q Con
 ctorForLeafType (DatatypeName name) typeParameterName = ctorForTypes name
-  [ ("ann",   varT typeParameterName) -- ann :: a
-  , ("bytes", conT ''Text)
+  [ ("ann",  varT typeParameterName) -- ann :: a
+  , ("text", conT ''Text)            -- text :: Text
   ]
 
 -- | Build Q Constructor for records
@@ -129,11 +125,13 @@ ctorForTypes constructorName types = recC (toName Named constructorName) recordF
 
 
 -- | Convert field types to Q types
-fieldTypesToNestedEither :: NonEmpty TreeSitter.Deserialize.Type -> Name -> Q TH.Type
-fieldTypesToNestedEither xs typeParameterName = foldr1 combine (fmap convertToQType xs) `appT` varT typeParameterName
+fieldTypesToNestedSum :: NonEmpty TreeSitter.Deserialize.Type -> Q TH.Type
+fieldTypesToNestedSum xs = go (toList xs)
   where
-    combine lhs rhs = (conT ''(:+:) `appT` lhs) `appT` rhs
+    combine lhs rhs = (conT ''(:+:) `appT` lhs) `appT` rhs -- (((((a :+: b) :+: c) :+: d)) :+: e)   ((a :+: b) :+: (c :+: d))
     convertToQType (MkType (DatatypeName n) named) = conT (toName named n)
+    go [x] = convertToQType x
+    go xs = let (l,r) = splitAt (length xs `div` 2) xs in (combine (go l) (go r))
 
 
 -- | Create bang required to build records
@@ -154,9 +152,20 @@ addTickIfNecessary s
 
 -- | Prepend "Anonymous" to named node when false, otherwise use regular toName
 toName :: Named -> String -> Name
-toName named str = mkName $ addTickIfNecessary $ case named of
+toName named str = mkName (toNameString named str)
+
+toNameString :: Named -> String -> String
+toNameString named str = addTickIfNecessary $ case named of
   Anonymous -> "Anonymous" <> toCamelCase str
   Named -> toCamelCase str
+
+-- | Get the 'Module', if any, for a given 'Name'.
+moduleForName :: Name -> Maybe Module
+moduleForName n = Module . PkgName <$> namePackage n <*> (ModName <$> nameModule n)
+
+-- | Test whether the name is defined in the module where the splice is executed.
+isLocalName :: Name -> Q Bool
+isLocalName n = (moduleForName n ==) . Just <$> thisModule
 
 -- Helper function to output camel cased data type names
 initUpper :: String -> String